% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_theme.R
\name{add_grid}
\alias{add_grid}
\title{Add gridlines}
\usage{
add_grid(color = "grey", linetype = "dotted", linewidth = 0.5, ...)
}
\description{
Add gridlines
}
\examples{
colors <- RColorBrewer::brewer.pal(n = 11, name = "BrBG")

iris \%>\%
  ggplot(aes(x = Sepal.Length, y = Sepal.Width, shape = Species, col = Species)) +
  geom_point(size = 2) +
  labs(title = "The famous iris data",
       subtitle = "Have you ever seen it?") +
  scale_color_manual(values = colors[c(8, 3, 10)]) +
  my_theme() +
  add_grid()
}
