% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc.R
\name{cc}
\alias{cc}
\title{Character vector without quotes
nomen est omeni}
\usage{
cc(...)
}
\arguments{
\item{...}{symbols to be passed to character}
}
\value{
character vector
}
\description{
Character vector without quotes
nomen est omeni
}
