% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notin.R
\name{\%notin\%}
\alias{\%notin\%}
\title{Not in}
\usage{
x \%notin\% y
}
\arguments{
\item{x}{vector}

\item{y}{vector}
}
\value{
logical
}
\description{
Inversion of \code{\%in\%}
}
\examples{
x <- c(1, 2, 3)
y <- c(2, 3, 4)
x \%in\% y
x \%notin\% y
}
