% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_theme.R
\name{legend_inside}
\alias{legend_inside}
\title{Place legend inside plot area (and scale)}
\usage{
legend_inside(
  position = c(1, 1),
  justification = c(1.1, 1.1),
  scale_title = 1,
  scale_text = 1,
  scale_key = 1,
  ...
)
}
\description{
Place legend inside plot area (and scale)
}
