% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swiss_shapefiles.R
\name{swiss_shapefiles}
\alias{swiss_shapefiles}
\title{Download and Read Swiss Shapefiles}
\usage{
swiss_shapefiles(url, readr = sf::st_read, ...)
}
\arguments{
\item{url}{from data.geo.admin.ch.}

\item{readr}{function to read shapefiles.}

\item{...}{additional arguments passed to \code{readr} func.}
}
\value{
List of shapefiles.
}
\description{
Download and Read Swiss Shapefiles
}
\examples{
swiss_shapefiels(url = "https://data.geo.admin.ch/ch.swisstopo.swissboundaries3d/swissboundaries3d_2024-01/swissboundaries3d_2024-01_2056_5728.shp.zip")
}
\seealso{
\link{https://data.geo.admin.ch/}
}
